refactor(react): use ark factory where it makes sense with simple components
