+++ packages/react/src/index.ts
@@ -15,6 +15,7 @@ export * from './components/deprecated/ModalDescription'
 export * from './components/deprecated/NavMenuTrigger'
 export * from './components/deprecated/NavMenuList'
 export * from './components/deprecated/NavMenuLink'
+export * from './components/deprecated/Legend'
 export * from './context/confirm-modal'
 export * from './context/prompt-modal'
 export * from './hooks/useDate'
@@ -41,7 +42,6 @@ export * from './components/FileStatus'
 export * from './components/FileUploader'
 export * from './components/for'
 export * from './components/IconButton'
-export * from './components/Legend'
 export * from './components/Menu'
 export * from './components/Notification'
 export * from './components/NotificationHeading'
